requirements.txt
cycler==0.10.0
dataclasses==0.8
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
